// src/types/pieces.ts
export enum pieces {
  KNIGHT_WHITE = "KNIGHT_WHITE",
  KNIGHT_DARK = "KNIGHT_DARK",
  ROOK_WHITE = "ROOK_WHITE",
  ROOK_DARK = "ROOK_DARK",
  KING_WHITE = "KING_WHITE",
  KING_DARK = "KING_DARK",
  QUEEN_WHITE = "QUEEN_WHITE",
  QUEEN_DARK = "QUEEN_DARK",
  PAWN_WHITE = "PAWN_WHITE",
  PAWN_DARK = "PAWN_DARK",
  BISHOP_WHITE = "BISHOP_WHITE",
  BISHOP_DARK = "BISHOP_DARK",
}